encoder-paecter/requirements.txt
numpy>=1.26,<2
psycopg2-binary>=2.9
transformers>=4.39
pyarrow>=15.0
